services/forecasting-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pandas==2.2.1
SQLAlchemy==2.0.28
psycopg2-binary==2.9.9
PyJWT==2.8.0
cryptography==42.0.5
python-dotenv==1.0.1
